move App component
