+++ src/store/routes.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import { Route } from 'react-router';
-import App from '../app/app.js';
+
+import App from '../components/App/App';
+
 export default (
   <Route path="/" component={ App }/>
 );
